+++ login-workflow/example/.gitignore
@@ -62,3 +62,6 @@ buck-out/
 # Ruby / CocoaPods
 /ios/Pods/
 /vendor/bundle/
+
+# Local Environment Config
+.env
