+++ login-workflow/example/babel.config.js
@@ -1,5 +1,5 @@
 module.exports = {
     presets: ['module:@react-native/babel-preset'],
     // react-native-reanimated/plugin must be listed last
-    plugins: ['react-native-reanimated/plugin'],
+    plugins: ['react-native-reanimated/plugin', 'module:react-native-dotenv', '@babel/plugin-transform-modules-commonjs',],
 };
