+++ login-workflow/src/screens/OktaLoginScreen/OktaLoginScreen.tsx
@@ -1,11 +1,8 @@
-import React, { useEffect } from 'react';
-import { createConfig, getAccessToken, signInWithBrowser, EventEmitter } from '@okta/okta-react-native';
+import React from 'react';
 import { OktaLoginScreenProps } from './types';
 import { OktaLoginScreenBase } from './OktaLoginScreenBase';
 import { useAuthContext } from '../../contexts';
-import { useErrorManager } from '../../contexts/ErrorContext/useErrorManager';
 import { useTranslation } from 'react-i18next';
-import oktaConfig from '../../okta.config';
 
 /**
  * Component that renders a okta login screen that prompts a user to redirect to okta login sign in page.
@@ -18,16 +15,7 @@ import oktaConfig from '../../okta.config';
 export const OktaLoginScreen: React.FC<React.PropsWithChildren<OktaLoginScreenProps>> = (props) => {
     const { t } = useTranslation();
     const auth = useAuthContext();
-    const { actions, navigate, routeConfig } = auth;
-    const { triggerError, errorManagerConfig } = useErrorManager();
-    const errorDisplayConfig = {
-        ...errorManagerConfig,
-        ...props.errorDisplayConfig,
-        onClose: (): void => {
-            if (props.errorDisplayConfig && props.errorDisplayConfig.onClose) props.errorDisplayConfig.onClose();
-            if (errorManagerConfig.onClose) errorManagerConfig?.onClose();
-        },
-    };
+    const { navigate, routeConfig } = auth;
 
     const {
         loginButtonLabel = t('bluiCommon:ACTIONS.OKTA_SIGN_IN'),
@@ -46,60 +34,13 @@ export const OktaLoginScreen: React.FC<React.PropsWithChildren<OktaLoginScreenPr
         header,
         footer,
         cyberSecurityBadgeSize,
-        oktaConfigObject,
+        onLogin,
     } = props;
-
-    const createOktaConfig = async (): Promise<void> => {
-        await createConfig(oktaConfigObject || oktaConfig);
-    };
-
-    useEffect(() => {
-        void actions.initiateSecurity();
-
-        EventEmitter.addListener('signInSuccess', () => {
-            // eslint-disable-next-line no-console
-            console.log('signInSuccess');
-        });
-        EventEmitter.addListener('signOutSuccess', () => {
-            // eslint-disable-next-line no-console
-            console.log('signOutSuccess');
-        });
-        EventEmitter.addListener('onError', (e: Event) => {
-            // eslint-disable-next-line no-console
-            console.log(e);
-        });
-        EventEmitter.addListener('onCancelled', (e: Event) => {
-            // eslint-disable-next-line no-console
-            console.log(e);
-        });
-
-        void createOktaConfig();
-
-        return () => {
-            EventEmitter.removeAllListeners('signInSuccess');
-            EventEmitter.removeAllListeners('signOutSuccess');
-            EventEmitter.removeAllListeners('onError');
-            EventEmitter.removeAllListeners('onCancelled');
-        };
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
 
     return (
         <OktaLoginScreenBase
             loginButtonLabel={loginButtonLabel}
-            onLogin={
-                (async (): Promise<void> => {
-                    try {
-                        await signInWithBrowser();
-                    } catch (_error) {
-                        triggerError(_error as Error);
-                    } finally {
-                        getAccessToken()// eslint-disable-next-line no-console
-                            .then((token) => console.log(token))// eslint-disable-next-line no-console
-                            .catch((error) => console.log('token error', error));
-                    }
-                }) as any
-            }
+            onLogin={onLogin}
             showForgotPassword={showForgotPassword}
             forgotPasswordLabel={forgotPasswordLabel}
             onForgotPassword={onForgotPassword}
@@ -110,7 +51,6 @@ export const OktaLoginScreen: React.FC<React.PropsWithChildren<OktaLoginScreenPr
             showContactSupport={showContactSupport}
             contactSupportLabel={contactSupportLabel}
             onContactSupport={onContactSupport}
-            errorDisplayConfig={errorDisplayConfig}
             showCyberSecurityBadge={showCyberSecurityBadge}
             projectImage={projectImage}
             header={header}
